Add Minicircuits RC SPDT to public api
