+++ qcodes/instrument_drivers/Minicircuits/__init__.py
@@ -1,8 +1,13 @@
 from ._minicircuits_rc_sp4t import MiniCircuitsRCSP4T, MiniCircuitsRCSP4TChannel
+from ._minicircuits_rc_spdt import MiniCircuitsRCSPDT, MiniCircuitsRCSPDTChannel
 from .RUDAT_13G_90 import MiniCircuitsRudat13G90Usb
 from .USB_SPDT import MiniCircuitsSwitchChannelUsb, MiniCircuitsUsbSPDT
 
 __all__ = [
+    "MiniCircuitsRCSP4T",
+    "MiniCircuitsRCSP4TChannel",
+    "MiniCircuitsRCSPDT",
+    "MiniCircuitsRCSPDTChannel",
     "MiniCircuitsRudat13G90Usb",
     "MiniCircuitsSwitchChannelUsb",
     "MiniCircuitsUsbSPDT",
